refactor: ICRUDDao.java, ICRUDService.java, IExecutorDao.java, IExecutorTaskDao.java, StatusCreateDTO.java, StatusDTO.java, TaskCreateDTO.java, TaskDTO.java, TaskJDBCDao.java, TaskService.java, TaskServiceFactory.java

+++ src/main/java/by/it_academy/jd2/Mk_JD2_98_23/dao/api/ICRUDDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface ICRUDDao<T> {
     List<T> get();
 
-    T get(int id);
+    T get(Long id);
 
     T save(T item);
